requirements.txt
autopep8==1.5.6
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
dnspython==1.16.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-PyMongo==2.3.0
Flask-Reuploaded==1.0.0
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Pillow==8.2.0
psycopg2==2.8.6
pycodestyle==2.7.0
pycparser==2.20
pymongo==3.11.3
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
six==1.15.0
timeago==1.0.15
toml==0.10.2
Werkzeug==1.0.1
